+++ frontend/src/lib/i18n.ts
@@ -7,8 +7,7 @@ import i18n from 'i18next'
 import { initReactI18next } from 'react-i18next'
 import LanguageDetector from 'i18next-browser-languagedetector'
 import HttpApi from 'i18next-http-backend'
-
-export const supportedLanguages = ['en', 'ja', 'ko', 'zh-CN']
+import { supportedLanguages } from './supportedLanguages'
 
 i18n
     // pass the i18n instance to react-i18next
