+++ frontend/src/lib/supportedLanguages.ts
@@ -1,0 +1,3 @@
+/* Copyright Contributors to the Open Cluster Management project */
+
+export const supportedLanguages = ['en', 'ja', 'ko', 'zh-CN']
